Migrate events.spec.ts from Cypress to Playwright
